[mlir][affine] Cancel exactly-matching delinearize/linearize pairs
If we linearize values (with an assertion tha they are disjoint) and
then delinearize that linear index with th exact same basis, we know
that these operations are exact inverses of each other and can be
replaced with the original inputs to the linearization.

Similarly, if we take a linear index, delinearize it with some bases,
and then re-linearize it with that same basis (noting that the outputs
of the delinearization are guaranteed to by `disjoint`, even if this
is not asserted on the linearize_index operation), the
re-linearization is the inverse of the delinearization, so those two
operations can also be canceled out.

This commit adds canonicalization patterns for these simple
cancelations.