[mlir][Arith] Let integer range narrowing handle negative values
Update integer range narrowing to handle negative values.

The previous restriction to only narrowing known-non-negative values
wasn't needed, as both the signed and unsigned ranges represent bounds
on the values of each variable in the program ... except that one
might be more accurate than the other. So, if either the signed or
unsigned interpretetation of the inputs and outputs allows for integer
narrowing, the narrowing is permitted.

This commit also updates the integer optimization rewrites to preserve
the stae of constant-like operations and those that are narrowed so
that rewrites of other operations don't lose that range information.